+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add buttons to toggle history/favorite suggestions
 - Add button to control if command is saved in history
 - Add button to toggle if result is open in a new editor
+- Add button when the selection is empty to toggle if all the text must be
+  selected
 - Persistent History
 - Environment variables can be predefined with `env.shell` settings 
 - Add `editWithShell.quickCommand` command to use in keybinding definition. It